Remove PostHook from EagerPR, always run for all pull-requests
Originally, the idea was users would need to enable this on either a
project or repository level if they needed to build pull-requests on
Bitbucket 7 or higher. The reasoning was it may have some performance
impact.

However, after many stress tests (lots of files, lots of commits,
running stress) this couldn't performance impact was neligable. Also
couldn't reproduce any sort of race-condition between the Internal PR
refresh & webhook firing. Updating the /from is almost instantaneous.
Effectivediff does take a few ms (which is where the possible race
conditrion was) but we don't actually care about the effectivediff, only
that /from has been updated

So we enable EagerPR on all projects & repositories.

Release 0.1.5